refactor(AST): separate modular AST transformer into traverser and transformer
This is a major refactoring of the structures hidden behind the create_modular_ast_transformer macro that is used to instantiate a Transformer that is at the same time also a Traverser. This commit splits the logic of traversing the AST and transforming it into two clearly separate types. For that purpose the new generic TraverseTransformer was born that traverses any given AST recursively and applies its internal transformer to all nodes recursively.