feat(AST): add TransformEvent to AST transformers
This enhances the entire Transformer trait and module to allow for finer grained simplification procedures.
For example the If-Constraint-Propagation simplification greatly benefits from this enhancement since it requires the additional tracking of entering and leaving of the current inspected AST node for proper management of the simplification steps.
This design also allows for another important enhancement: Post-Processing.
Simplification procedures such as ITE-Lifting (from STP) require a post-processing step to finalize their simplifications after all AST transformations have taken place in an ordinary fashion.

The chosen design to extend the Transformer trait does not break any existing code but enables the enhancements at the same time.