refactor(AST): move to proper error management using custom error types
This is a major refactoring and finally allows for proper error handling and error management. It was decided to _NOT_ use the new failure crate since that is better be used by binary crates and not by library crates. This also allows a self-standing type- and expression-validation system and other neat things.