feat(AST): add experimental initial implementation of Context
This will replace the current global static SymbolInterner and the global static TypeMap found in symbol.rs.
Those are known to create problems especially for testability of the crate.
Some people have recommended the rental crate for help with self-referencing structs and lifetime corrections as can be found in the resolve method of the SymbolInterner, however, those add memory and operational overhead to the scene and with the current invariants the system is sound and correct as it is but may remain a maintainability pitfall with the current implementation.