refactor(Parser/Lexer): add simple_lexer, rename lexer to raw_lexer
This distinction is important and affects the resulting tokens and their kinds.
The raw tokens are easily and more efficiently scanned but the simple tokens are easier for the following parsing step to handle and thus used.
The simple lexer reads the raw token stream of its internal raw lexer and converts the raw tokens into simple tokens.